feat(layout): Created contacts manager page, but layout isnt there yet
Im having a hard time getting a vertical layout kind of like Protonmail. I want the columns to take up the whole width of the page, but it only goes as far as the content in the list-group. I want it so if theres not enough list items in the list-group, then itll show a background where list items would go.