feat(layout): I finally fucking got this shit to be layed out how I want it too HEEEELLLLLYEAH.
The columns now properly fill the whole page, but that can be modified to be closer like the UX design, but Im thinking of changing the design to be something close to how ProtonMail is laid out. Anyway, to modify the layout to be like the original design, change the height for body, html, and main in manager.css, and change the main class from "container-fluid" to "container" or "container-*".