feat(form): Added edit, save, cancel, delete buttons to form
Delete doesn't work yet, but cancel and save do locally. No API endpoints are integrated, but the code is in the position to accept work on them now. Need to add alerts for failing to select a different contact while there are changes on the form (currently just blocks you from selecting it). Also, there will need to be alerts for successful saves/deletes w/ server messages.